requirements.txt
pandas
geopandas
fiona
tqdm
shapely
matplotlib
xarray
rasterio
numpy
rioxarray
dask
distributed
pystac-client
planetary-computer
odc-stac